+++ src/biz/bokhorst/xprivacy/PrivacyService.java
@@ -67,7 +67,7 @@ public class PrivacyService {
 	private static final String cTableUsage = "usage";
 	private static final String cTableSetting = "setting";
 
-	private static final int cCurrentVersion = 267;
+	private static final int cCurrentVersion = 268;
 	private static final String cServiceName = "xprivacy" + cCurrentVersion;
 
 	// TODO: define column names