+++ CHANGELOG.md
@@ -44,6 +44,8 @@ Test and beta releases will have experimental functions enabled by default.
 
 **Next release**
 
+* Fixed all issues reported with the support info
+
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 1.99.33-3 TEST**