+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Test and beta releases will have experimental functions enabled by default.
 **Next release**
 
 * Fixed all issues reported with the support info
+* Updated Polish translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)